pyproject.toml
[project]
name = "aiko-gpt"
version = "0.1.0"
description = "Générateur de rapports assisté par IA"
authors = [
    {name = "AIKO", email = "contact@aiko.fr"}
]
requires-python = ">=3.9"
dependencies = [
    "pandas>=2.0.0",
    "openpyxl>=3.1.0",
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "streamlit>=1.28.0",
    "langgraph[checkpoint]>=0.4.0",
    "langchain>=0.1.0",
    "langchain-openai>=0.0.5",
    "pdfplumber>=0.10.0",
    "langgraph-cli[inmem]>=0.4.3",
    "langgraph-checkpoint-sqlite>=2.0.11",
    "python-docx>=1.0.0",
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "python-multipart>=0.0.6",
    "perplexityai>=0.16.0",
    "sqlalchemy>=2.0.0",
    "psycopg2-binary>=2.9.0",
    "alembic>=1.13.0",
]

[tool.uv]
dev-dependencies = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
]
